third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/stroke-width-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'stroke-width' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
[FAIL] Can set 'stroke-width' to a number: 0
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-width' to a number: -3.14
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-width' to a number: 3.14
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'stroke-width' to a number: calc(2 + 3)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
Harness: the test ran to completion.
